+++ R/tm_a_mmrm.R
@@ -809,15 +809,20 @@ srv_mmrm <- function(input,
 
   anl_merged <- data_merge_module(
     datasets = datasets,
-    data_extract = list(arm_var, paramcd, id_var, visit_var, split_covariates, aval_var),
-    input_id = c("arm_var", "paramcd", "id_var", "visit_var", "split_covariates", "aval_var"),
+    data_extract = list(
+      arm_var = arm_var,
+      paramcd = paramcd,
+      id_var = id_var,
+      visit_var = visit_var,
+      split_covariates = split_covariates,
+      aval_var = aval_var
+    ),
     merge_function = "dplyr::inner_join"
   )
 
   adsl_merged <- data_merge_module(
     datasets = datasets,
-    data_extract = list(arm_var),
-    input_id = c("arm_var"),
+    data_extract = list(arm_var = arm_var),
     anl_name = "ANL_ADSL"
   )