+++ R/tm_t_abnormality.R
@@ -475,21 +475,29 @@ srv_t_abnormality <- function(input,
     choices <- unique(anl[[input$grade]][!is.na(anl[[input$grade]])])
   })
 
-  anl_merged <- data_merge_module(
+  anl_selectors <- data_extract_multiple_srv(
+    list(
+      arm_var = arm_var,
+      id_var = id_var,
+      by_vars = by_vars,
+      grade = grade,
+      baseline_var = baseline_var,
+      treatment_flag_var = treatment_flag_var
+    ),
+    datasets = datasets
+  )
+
+  anl_merged <- data_merge_srv(
+    selector_list = anl_selectors,
     datasets = datasets,
-    data_extract = list(arm_var, id_var, by_vars, grade, baseline_var, treatment_flag_var),
-    input_id = c("arm_var", "id_var", "by_vars", "grade", "baseline_var", "treatment_flag_var"),
     merge_function = "dplyr::inner_join"
   )
 
   adsl_merged <- data_merge_module(
     datasets = datasets,
-    data_extract = list(arm_var),
-    input_id = c("arm_var"),
+    data_extract = list(arm_var = arm_var),
     anl_name = "ANL_ADSL"
   )
-
-  by_vars_ordered <- get_input_order("by_vars", by_vars$dataname)
 
   validate_checks <- reactive({
     adsl_filtered <- datasets$get_data(parentname, filtered = TRUE)
@@ -498,7 +506,7 @@ srv_t_abnormality <- function(input,
     anl_m <- anl_merged()
     input_arm_var <- as.vector(anl_m$columns_source$arm_var)
     input_id_var <- as.vector(anl_m$columns_source$id_var)
-    input_by_vars <- by_vars_ordered()
+    input_by_vars <- anl_selectors()$by_vars()$select_ordered
     input_grade <- as.vector(anl_m$columns_source$grade)
     input_baseline_var <- as.vector(anl_m$columns_source$baseline_var)
     input_treatment_flag_var <- as.vector(anl_m$columns_source$treatment_flag_var)
@@ -537,8 +545,8 @@ srv_t_abnormality <- function(input,
       parentname = "ANL_ADSL",
       dataname = "ANL",
       arm_var = as.vector(anl_m$columns_source$arm_var),
+      by_vars = anl_selectors()$by_vars()$select_ordered,
       id_var = as.vector(anl_m$columns_source$id_var),
-      by_vars = by_vars_ordered(),
       abnormal = abnormal,
       grade = as.vector(anl_m$columns_source$grade),
       baseline_var = as.vector(anl_m$columns_source$baseline_var),