+++ R/tm_t_tte.R
@@ -710,15 +710,20 @@ srv_t_tte <- function(input,
 
   anl_merged <- data_merge_module(
     datasets = datasets,
-    data_extract = list(arm_var, paramcd, aval_var, cnsr_var, strata_var, event_desc_var, time_unit_var),
-    input_id = c("arm_var", "paramcd", "aval_var", "cnsr_var", "strata_var", "event_desc_var", "time_unit_var"),
+    data_extract = list(
+      arm_var = arm_var,
+      paramcd = paramcd,
+      aval_var = aval_var,
+      cnsr_var = cnsr_var,
+      strata_var = strata_var,
+      event_desc_var = event_desc_var,
+      time_unit_var = time_unit_var),
     merge_function = "dplyr::inner_join"
   )
 
   adsl_merged <- data_merge_module(
     datasets = datasets,
-    data_extract = list(arm_var, strata_var),
-    input_id = c("arm_var", "strata_var"),
+    data_extract = list(arm_var = arm_var, strata_var = strata_var),
     anl_name = "ANL_ADSL"
   )
 
@@ -808,7 +813,7 @@ srv_t_tte <- function(input,
       dataname = "ANL",
       parentname = "ANL_ADSL",
       arm_var = as.vector(anl_m$columns_source$arm_var),
-      paramcd = unlist(anl_m$filter_info)["selected"],
+      paramcd = unlist(anl_m$filter_info$paramcd)["selected"],
       ref_arm = input$ref_arm,
       comp_arm = input$comp_arm,
       compare_arm = input$compare_arms,