+++ tests/testthat/test-tm_t_rsp.R
@@ -18,8 +18,8 @@ test_that("template_rsp generates standard expressions", {
         dplyr::filter(ARMCD %in% c("ARM A", "ARM B", "ARM C")) %>%
         dplyr::mutate(ARMCD = stats::relevel(ARMCD, ref = "ARM A")) %>%
         dplyr::mutate(ARMCD = droplevels(ARMCD)) %>%
-        dplyr::mutate(is_rsp = AVALC %in% c("CR", "PR")) %>%
-        df_explicit_na()
+        dplyr::mutate(is_rsp = AVALC %in% c("Complete Response (CR)", "Partial Response (PR)")) %>%
+        dplyr::mutate(AVALC = factor(AVALC, levels = c("Complete Response (CR)", "Partial Response (PR)")))
       adsl <- adsl %>%
         dplyr::filter(ARMCD %in% c("ARM A", "ARM B", "ARM C")) %>%
         dplyr::mutate(ARMCD = stats::relevel(ARMCD, ref = "ARM A")) %>%
@@ -29,9 +29,9 @@ test_that("template_rsp generates standard expressions", {
     layout = quote(
       lyt <- basic_table(
         title = paste("Table of", "BESRSPI",
-                      "for", paste(head(c("CR", "PR"), -1), collapse = ", "),
-                      ifelse(length(c("CR", "PR")) > 1, "and",
-                             ""), tail(c("CR", "PR"), 1), "Responders")
+                      "for", paste(head(c("Complete Response (CR)", "Partial Response (PR)"), -1), collapse = ", "),
+                      ifelse(length(c("Complete Response (CR)", "Partial Response (PR)")) > 1, "and",
+                             ""), tail(c("Complete Response (CR)", "Partial Response (PR)"), 1), "Responders")
       ) %>%
         split_cols_by(var = "ARMCD", ref_group = "ARM A") %>%
         add_colcounts() %>%
@@ -85,7 +85,8 @@ test_that("template_rsp generates right expressions with non-default", {
     comp_arm = c("A: Drug X", "C: Combination"),
     compare_arm = TRUE,
     combine_comp_arms = FALSE,
-    show_rsp_cat = FALSE
+    show_rsp_cat = FALSE,
+    responder_val = c("PR", "SD")
   )
 
   expected <- list(
@@ -94,8 +95,8 @@ test_that("template_rsp generates right expressions with non-default", {
         dplyr::filter(ARM %in% c("B: Placebo", "A: Drug X", "C: Combination")) %>%
         dplyr::mutate(ARM = stats::relevel(ARM, ref = "B: Placebo")) %>%
         dplyr::mutate(ARM = droplevels(ARM)) %>%
-        dplyr::mutate(is_rsp = AVALC %in% c("CR", "PR")) %>%
-        df_explicit_na()
+        dplyr::mutate(is_rsp = AVALC %in% c("PR", "SD")) %>%
+        dplyr::mutate(AVALC = factor(AVALC, levels = c("PR", "SD")))
       ADSL <- ADSL %>% # nolint
         dplyr::filter(ARM %in% c("B: Placebo", "A: Drug X", "C: Combination")) %>%
         dplyr::mutate(ARM = stats::relevel(ARM, ref = "B: Placebo")) %>%
@@ -105,9 +106,9 @@ test_that("template_rsp generates right expressions with non-default", {
     layout = quote(
       lyt <- basic_table(
         title = paste("Table of", "BESRSPI", "for",
-                      paste(head(c("CR", "PR"), -1), collapse = ", "),
-                      ifelse(length(c("CR", "PR")) > 1,
-                             "and", ""), tail(c("CR", "PR"), 1), "Responders")
+                      paste(head(c("PR", "SD"), -1), collapse = ", "),
+                      ifelse(length(c("PR", "SD")) > 1,
+                             "and", ""), tail(c("PR", "SD"), 1), "Responders")
         ) %>%
         split_cols_by(var = "ARM", ref_group = "B: Placebo") %>%
         add_colcounts() %>%
@@ -162,8 +163,8 @@ test_that("template_rsp generates expression without arm comparison", {
     data = quote({
       anl <- ADRS %>%
         dplyr::mutate(ARM = droplevels(ARM)) %>%
-        dplyr::mutate(is_rsp = AVALC %in% c("CR", "PR")) %>%
-        df_explicit_na()
+        dplyr::mutate(is_rsp = AVALC %in% c("Complete Response (CR)", "Partial Response (PR)")) %>%
+        dplyr::mutate(AVALC = factor(AVALC, levels = c("Complete Response (CR)", "Partial Response (PR)")))
       ADSL <- ADSL %>%  # nolint
         dplyr::mutate(ARM = droplevels(ARM)) %>%
         df_explicit_na()
@@ -171,9 +172,9 @@ test_that("template_rsp generates expression without arm comparison", {
     layout = quote(
       lyt <- basic_table(
         title = paste("Table of", "BESRSPI",
-                      "for", paste(head(c("CR", "PR"), -1), collapse = ", "),
-                      ifelse(length(c("CR", "PR")) > 1, "and",
-                             ""), tail(c("CR", "PR"), 1), "Responders")
+                      "for", paste(head(c("Complete Response (CR)", "Partial Response (PR)"), -1), collapse = ", "),
+                      ifelse(length(c("Complete Response (CR)", "Partial Response (PR)")) > 1, "and",
+                             ""), tail(c("Complete Response (CR)", "Partial Response (PR)"), 1), "Responders")
         ) %>%
         split_cols_by(var = "ARM") %>%
         add_colcounts() %>%
@@ -219,8 +220,8 @@ test_that("template_rsp generates expression with non-default controls and strat
         dplyr::filter(ARM %in% c("B: Placebo", "A: Drug X", "C: Combination")) %>%
         dplyr::mutate(ARM = stats::relevel(ARM, ref = "B: Placebo")) %>%
         dplyr::mutate(ARM = droplevels(ARM)) %>%
-        dplyr::mutate(is_rsp = AVALC %in% c("CR", "PR")) %>%
-        df_explicit_na()
+        dplyr::mutate(is_rsp = AVALC %in% c("Complete Response (CR)", "Partial Response (PR)")) %>%
+        dplyr::mutate(AVALC = factor(AVALC, levels = c("Complete Response (CR)", "Partial Response (PR)")))
       ADSL <- ADSL %>% # nolint
         dplyr::filter(ARM %in% c("B: Placebo", "A: Drug X", "C: Combination")) %>% #nolint
         dplyr::mutate(ARM = stats::relevel(ARM, ref = "B: Placebo")) %>%
@@ -230,9 +231,9 @@ test_that("template_rsp generates expression with non-default controls and strat
     layout = quote(
       lyt <- basic_table(
         title = paste("Table of", "BESRSPI", "for",
-                      paste(head(c("CR", "PR"), -1), collapse = ", "),
-                      ifelse(length(c("CR", "PR")) > 1, "and",
-                             ""), tail(c("CR", "PR"), 1), "Responders")
+                      paste(head(c("Complete Response (CR)", "Partial Response (PR)"), -1), collapse = ", "),
+                      ifelse(length(c("Complete Response (CR)", "Partial Response (PR)")) > 1, "and",
+                             ""), tail(c("Complete Response (CR)", "Partial Response (PR)"), 1), "Responders")
         ) %>%
         split_cols_by(var = "ARM", ref_group = "B: Placebo") %>%
         add_colcounts() %>%
@@ -311,8 +312,8 @@ test_that("template_rsp can combine comparison arms", {
         dplyr::filter(ARMCD %in% c("ARM A", "ARM B", "ARM C")) %>%
         dplyr::mutate(ARMCD = stats::relevel(ARMCD, ref = "ARM A")) %>%
         dplyr::mutate(ARMCD = droplevels(ARMCD)) %>%
-        dplyr::mutate(is_rsp = AVALC %in% c("CR", "PR")) %>%
-        df_explicit_na()
+        dplyr::mutate(is_rsp = AVALC %in% c("Complete Response (CR)", "Partial Response (PR)")) %>%
+        dplyr::mutate(AVALC = factor(AVALC, levels = c("Complete Response (CR)", "Partial Response (PR)")))
       ADSL <- ADSL %>% # nolint
         dplyr::filter(ARMCD %in% c("ARM A", "ARM B", "ARM C")) %>%
         dplyr::mutate(ARMCD = stats::relevel(ARMCD, ref = "ARM A")) %>%
@@ -325,9 +326,9 @@ test_that("template_rsp can combine comparison arms", {
     layout = quote(
       lyt <- basic_table(
         title = paste("Table of", "BESRSPI", "for",
-                      paste(head(c("CR", "PR"), -1), collapse = ", "),
-                      ifelse(length(c("CR", "PR")) > 1, "and",
-                             ""), tail(c("CR", "PR"), 1), "Responders")
+                      paste(head(c("Complete Response (CR)", "Partial Response (PR)"), -1), collapse = ", "),
+                      ifelse(length(c("Complete Response (CR)", "Partial Response (PR)")) > 1, "and",
+                             ""), tail(c("Complete Response (CR)", "Partial Response (PR)"), 1), "Responders")
         ) %>%
         split_cols_by_groups(
           var = "ARMCD", groups_list = groups, ref_group = names(groups)[1]
@@ -406,8 +407,8 @@ test_that("template_rsp can combine refs", {
         dplyr::mutate(ARMCD = combine_levels(ARMCD, levels = c("ARM A", "ARM B"), new_level = "ARM A/ARM B")) %>%
         dplyr::mutate(ARMCD = stats::relevel(ARMCD, ref = "ARM A/ARM B")) %>%
         dplyr::mutate(ARMCD = droplevels(ARMCD)) %>%
-        dplyr::mutate(is_rsp = AVALC %in% c("CR", "PR")) %>%
-        df_explicit_na()
+        dplyr::mutate(is_rsp = AVALC %in% c("Complete Response (CR)", "Partial Response (PR)")) %>%
+        dplyr::mutate(AVALC = factor(AVALC, levels = c("Complete Response (CR)", "Partial Response (PR)")))
       adsl <- adsl %>%
         dplyr::filter(ARMCD %in% c("ARM A", "ARM B", "ARM C")) %>%
         dplyr::mutate(ARMCD = combine_levels(ARMCD, levels = c("ARM A", "ARM B"), new_level = "ARM A/ARM B")) %>%
@@ -418,9 +419,9 @@ test_that("template_rsp can combine refs", {
     layout = quote(
       lyt <- basic_table(
         title = paste("Table of", "BESRSPI", "for",
-                      paste(head(c("CR", "PR"), -1), collapse = ", "),
-                      ifelse(length(c("CR", "PR")) > 1, "and",
-                             ""), tail(c("CR", "PR"), 1), "Responders")
+                      paste(head(c("Complete Response (CR)", "Partial Response (PR)"), -1), collapse = ", "),
+                      ifelse(length(c("Complete Response (CR)", "Partial Response (PR)")) > 1, "and",
+                             ""), tail(c("Complete Response (CR)", "Partial Response (PR)"), 1), "Responders")
         ) %>%
         split_cols_by(var = "ARMCD", ref_group = "ARM A/ARM B") %>%
         add_colcounts() %>%