+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.modules.clinical 0.8.10.9018
+# teal.modules.clinical 0.8.10.9023
 
 ### Breaking changes
 * Updated `tm_t_abnormality` due to changes in `count_abnormal` that `abnormal` argument is taking list as input now.
@@ -8,14 +8,18 @@
 * Enhanced `tm_t_ancova` to include selected parameter(s), visit(s) and the analysis variable in title of the table.
 * New module `tm_g_lineplot` was added for creating line plots.
 * Enhanced `tm_t_logistic` to include selected parameter in title of the table.
-
+* Enhanced `tm_g_forest_rsp` to include selected parameter in title of the table. 
+* Enhanced `tm_g_forest_tte` to include selected parameter in title of the table. 
 
 ## Enhancements
 * Added support for logging with the `logger` package and added info level logs upon initialization of a module.
+* Added `default_responses` argument to `tm_t_binary_outcome`, `tm_t_rsp`, `tm_g_forest_rsp` to allow the user to specify default selected responses and possible response levels.
+* Updated `tm_t_binary_outcome` to show only the selected responses in the output table when "Show All Selected Response Categories" is checked.
 
 ### Miscellaneous
 * Updated R version requirement to >= 3.6.
 * Refactored calls to the defunct `teal.devel::data_extract_input` into calls to its replacement `teal.devel::data_extract_ui`.
+* Updated modules to use new `data_merge_module` interface provided by teal.devel and removed usage of the now deprecated function `teal.devel::get_input_order`.
 
 # teal.modules.clinical 0.8.10
 ### New features